@@ -32,22 +32,22 @@
     "@types/chai": "^4.3.5",
     "@types/express": "^4.17.17",
     "@types/mocha": "^10.0.1",
-    "@types/node": "^20.2.5",
+    "@types/node": "^20.8.0",
     "@types/pg": "^8.10.1",
-    "@typescript-eslint/eslint-plugin": "^5.59.7",
+    "@typescript-eslint/eslint-plugin": "^6.7.3",
     "chai": "^4.3.7",
     "chai-http": "^4.3.0",
     "concurrently": "^8.0.1",
-    "eslint": "^8.41.0",
+    "eslint": "^8.50.0",
     "eslint-config-custom": "*",
-    "eslint-config-standard-with-typescript": "^34.0.1",
+    "eslint-config-standard-with-typescript": "^39.1.0",
     "eslint-plugin-import": "^2.27.5",
-    "eslint-plugin-n": "^15.7.0",
-    "eslint-plugin-prettier": "^4.2.1",
+    "eslint-plugin-n": "^16.1.0",
+    "eslint-plugin-prettier": "^5.0.0",
     "eslint-plugin-promise": "^6.1.1",
     "mocha": "^10.2.0",
-    "nodemon": "^2.0.22",
-    "prettier": "2.8.8",
+    "nodemon": "^3.0.1",
+    "prettier": "3.0.3",
     "rimraf": "^5.0.1",
     "tsconfig": "workspace:*",
     "typescript": "^5.0.4"
