@@ -11,11 +11,11 @@
   },
   "devDependencies": {
     "@turbo/gen": "^1.9.9",
-    "eslint": "^7.32.0",
+    "eslint": "^8.50.0",
     "eslint-config-custom": "workspace:*",
     "husky": "^8.0.0",
     "lint-staged": "^14.0.1",
-    "prettier": "^2.8.8",
+    "prettier": "^3.0.3",
     "turbo": "^1.9.9"
   },
   "packageManager": "pnpm@7.15.0",
