@@ -4,11 +4,11 @@
   "main": "index.js",
   "license": "MIT",
   "dependencies": {
-    "eslint": "^7.32.0",
-    "eslint-config-next": "^13.4.4",
-    "eslint-config-prettier": "^8.8.0",
+    "eslint": "^8.50.0",
+    "eslint-config-next": "^13.5.4",
+    "eslint-config-prettier": "^9.0.0",
     "eslint-config-turbo": "^1.9.9",
-    "eslint-plugin-react": "7.28.0"
+    "eslint-plugin-react": "7.33.2"
   },
   "devDependencies": {
     "tsconfig": "workspace:*"
