@@ -1,8 +1,7 @@
 import { Request, Response, NextFunction } from 'express';
+
 const jwt = require('jsonwebtoken');
 const pg = require('pgdatabase').pg;
-const Organization = require('common').Organization;
-const Event = require('common').Event;
 const User = require('common').User;
 
 const accessTokenSecret = process.env.SECRET_KEY;
@@ -12,7 +11,7 @@ const authorize = async (req: Request & typeof User, res: Response, next: NextFu
   const authHeader = req.headers.authorization;
   if (authHeader) {
     const token = authHeader.split(' ')[1];
-    jwt.verify(token, accessTokenSecret, (err: Error, user: typeof User) => {
+    jwt.verify(token, accessTokenSecret, async (err: Error, user: typeof User) => {
       if (err) {
         if (err.name === 'TokenExpiredError') {
           return res.status(401).json({ error: 'Token expired' });
@@ -20,7 +19,15 @@ const authorize = async (req: Request & typeof User, res: Response, next: NextFu
         return res.sendStatus(403);
       }
 
+      const organizations = (
+        await pg.query(
+          'SELECT * FROM ORGANIZATION JOIN ORGANIZATION_USER ON ORGANIZATION.ID = ORGANIZATION_USER.ORGANIZATION_ID WHERE ORGANIZATION_USER.USER_ID = $1',
+          [user.id],
+        )
+      )?.rows;
+
       req.body.user = user;
+      req.body.organizations = organizations.rows;
       next();
     });
   } else {
